requirements.txt
fastapi
uvicorn[standard]
gunicorn
assemblyai
python-dotenv
pydantic